app/requirements.txt
Django==2.2.5
psycopg2-binary==2.8.3
djangorestframework==3.10.3
PyJWT==1.7.1
djangorestframework-simplejwt==4.3.0